Remove cases importing react-impulse from impulse-form tests
